+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fixed crash in storage cache (raoulvdberge)
 - Fixed Crafting Pattern model (pauljoda)
 - Fixed Constructor not working on Botania flowers (raoulvdberge)
+- Fixed Disk Manipulator crash (raoulvdberge)
 
 ### 1.2.22
 - Fixed recipe for Processing Pattern Encoder not using oredictionary for the workbench (VT-14)